requirements.txt

pandas 
requests 
sqlalchemy
google-cloud-storage
python-dotenv
apache-airflow
gcsfs